Add GHC 9.10 support
Bumping the dependency bounds for `base` and `template-haskell` is
sufficient.  There are no changes to the code.

Bumping the same dependency bounds for `explainable-predicates` is also
required.  This commit was tested locally using `allow-newer`, but it
will fail CI until the `explainable-predicates` update is released to
Hackage.

I noticed that the README was out-of-date, so this commit updates the
README as well as CHANGELOG.  Please note that CHANGELOG update is in
a new "Unreleased" section that needs to be updated the with the
version and date upon release.